+++ IDE/src/main/java/org/sikuli/ide/PatternSimilaritySlider.java
@@ -44,7 +44,7 @@ class PatternSimilaritySlider extends JSlider {
 	}
 
 	public void showValue(JLabel lbl, int val) {
-		float sim = val > 99 ? 0.99f : (float) val / 100;
+		double sim = val > 99 ? 0.99 : (double) val / 100;
 		String txt = String.format("   ( %.2f )", sim);
 		lbl.setText(txt);
 		lbl.repaint();
